+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Plotpot changelog
 =================
 
+v1.0.1 - 2017-08-15
+-------------------
+
+-  fix smooth option
+-  improve the help output
+-  renamed counter electrode option into --biologic_ce
+
 v1.0.0 - 2017-08-13
 -------------------